requirements.txt
asgiref==3.8.1
Django==5.0.6
django-adminlte-ui==2.1.1
django-ckeditor==6.7.1
django-jazzmin==2.5.0
django-js-asset==2.2.0
django-modeltranslation==0.19.7
django-suit-v2==2.0
pillow==10.4.0
psycopg2==2.9.9
python-gettext==5.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0
